updated requirements.in and ensemble tide modelling for exposure (#38)
* updated requirements.in and ensemble tide modelling for exposure

* Rearranged exposure.py to move active function to top of the script
