+++ requirements.in
@@ -4,7 +4,7 @@ bokeh==3.2.2  # temporary fix for Dask issue
 botocore
 click==8.1.3
 datacube[s3,performance]==1.8.13
-dea-tools==0.3.1.dev1
+dea-tools==0.3.1.dev5
 geopandas==0.13.2
 matplotlib==3.7.1
 numpy==1.24.3
@@ -18,6 +18,7 @@ rasterio==1.3.4
 scikit_image==0.19.3
 scikit_learn==1.2.2
 scipy==1.10.1
+sunriset==1.0
 Shapely==2.0.1
 tqdm==4.65.0
 xarray==2023.1.0
